# api/openapi.yaml
openapi: 3.0.3
info:
  title: Axiom owner-console release API
  version: 1.3.0
  description: >-
    Authenticated research, operations, audit, sandbox, and qualification
    control plane. Axiom V1 is Spot-only research software. Real trading,
    production-private order submission, and external asset movement are absent.
servers:
  - url: /
...
paths:
...
  /api/v1/risk/pause:
    post:
      operationId: pauseRisk
      tags: [Risk]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/RevisionCommand"
      responses:
        "202":
          description: Durable pause command accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandAccepted" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
  /api/v1/risk/resume:
    post:
      operationId: resumeRisk
      tags: [Risk]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/RevisionCommand"
      responses:
        "202":
          description: Policy-gated durable resume command accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandAccepted" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /api/v1/backtests:
    post:
      operationId: createBacktest
      tags: [Backtest]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OfflineJobRequest" }
      responses:
        "202":
          description: Durable backtest job accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobResource" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /api/v1/replays:
    post:
      operationId: createReplay
      tags: [Replay]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReplayJobRequest" }
      responses:
        "202":
          description: Durable replay job accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobResource" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/replays/{id}:
    get:
      operationId: getReplay
      tags: [Replay]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - in: query
          name: event_ordinal
          description: Select one persisted canonical replay event; defaults to the newest available event.
          schema: { $ref: "#/components/schemas/Revision" }
      responses:
        "200":
          description: Replay lifecycle and deterministic cursor projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobResource" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/shadow-sessions:
    get:
      operationId: listShadowSessions
      tags: [Shadow]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: state
          schema:
            type: string
            enum: [QUEUED, RUNNING, PAUSED, CANCEL_REQUESTED, CANCELED, FAILED]
      responses:
        "200":
          description: Bounded deterministic public-data shadow history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShadowSessionPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createShadowSession
      tags: [Shadow]
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ShadowSessionRequest" }
      responses:
        "202":
          description: Public-data simulation-only shadow session accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShadowSessionResource" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/shadow-sessions/{id}/stop:
    post:
      operationId: stopShadowSession
      tags: [Shadow]
      security: [{ sessionCookie: [] }]
      parameters: *replayCommandParameters
      requestBody:
        $ref: "#/components/requestBodies/RevisionCommand"
      responses: *commandResponses
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: axiom_session
    operationalBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Origin:
      in: header
      name: Origin
      required: true
      schema: { type: string, format: uri, maxLength: 512 }
    CSRFToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string, minLength: 32, maxLength: 512 }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        {
          type: string,
          minLength: 16,
          maxLength: 128,
          pattern: "^[A-Za-z0-9._:-]+$",
        }
    ResourceID:
      in: path
      name: id
      required: true
      schema: { type: string, minLength: 1, maxLength: 160 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string, minLength: 16, maxLength: 1024 }
    PageSize:
      in: query
      name: page_size
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  requestBodies:
    RevisionCommand:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RevisionCommandRequest" }
  responses:
    BadRequest:
      description: Request failed safe validation
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Authentication failed
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Authorization, Origin, or CSRF policy denied the request
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Idempotency payload or lifecycle conflict
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    PreconditionFailed:
      description: Expected entity revision is stale
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    RateLimited:
      description: Durable authentication or workload quota exceeded
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unavailable:
      description: A required fail-closed prerequisite is unavailable
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Decimal:
      type: string
      pattern: "^-?(0|[1-9][0-9]*)(\\.[0-9]+)?$"
      example: "1000.000000000000000000"
    NonnegativeDecimal:
      type: string
      pattern: "^(0|[1-9][0-9]*)(\\.[0-9]+)?$"
    Revision:
      type: string
      pattern: "^(0|[1-9][0-9]*)$"
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp with an explicit UTC offset.
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlation_id]
      properties:
        code: { type: string, minLength: 1, maxLength: 80 }
        message: { type: string, minLength: 1, maxLength: 240 }
        correlation_id: { type: string, minLength: 1, maxLength: 160 }
        summary: { type: string, minLength: 1, maxLength: 240 }
        detail: { type: string, minLength: 1, maxLength: 2000 }
        impact: { type: string, minLength: 1, maxLength: 500 }
        suggested_action: { type: string, minLength: 1, maxLength: 500 }
        blocking_prerequisites:
          type: array
          maxItems: 16
          items: { type: string, minLength: 1, maxLength: 240 }
        current_state: { type: string, minLength: 1, maxLength: 160 }
        required_state: { type: string, minLength: 1, maxLength: 160 }
        field_details:
          type: object
          additionalProperties: { type: string, maxLength: 160 }
    Page:
      type: object
      additionalProperties: false
      required: [revision, has_more]
      properties:
        revision: { $ref: "#/components/schemas/Revision" }
        next_cursor: { type: string, minLength: 16, maxLength: 1024 }
        has_more: { type: boolean }
...
    RevisionCommandRequest:
      type: object
      additionalProperties: false
      required: [expected_revision, reason]
      properties:
        expected_revision: { $ref: "#/components/schemas/Revision" }
        reason: { type: string, minLength: 8, maxLength: 500 }
    CommandAccepted:
      type: object
      additionalProperties: false
      required: [id, state, target_id, revision, correlation_id, created_at]
      properties:
        id: { type: string }
        state: { type: string, enum: [pending, applied, rejected, failed] }
        target_id: { type: string }
        revision: { $ref: "#/components/schemas/Revision" }
        correlation_id: { type: string }
        created_at: { $ref: "#/components/schemas/Timestamp" }
...
    OfflineJobRequest:
      type: object
      additionalProperties: false
      required:
        [
          dataset_id,
          strategy_version,
          configuration_id,
          research_generation_id,
          root_seed_hash,
        ]
      properties:
        dataset_id: { type: string }
        strategy_version:
          {
            type: string,
            enum:
              [
                trend-following@1.0.0,
                mean-reversion@1.0.0,
                triangular-arbitrage@1.0.0,
                cross-exchange-arbitrage@1.0.0,
                inventory-rebalancing@1.0.0,
              ],
          }
        configuration_id: { type: string }
        research_generation_id: { type: string }
        root_seed_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
    ReplayJobRequest:
      allOf:
        - $ref: "#/components/schemas/OfflineJobRequest"
        - type: object
          properties:
            speed: { type: string, enum: [original, accelerated, maximum] }
            incident_id: { type: string }
            first_ordinal: { $ref: "#/components/schemas/Revision" }
            last_ordinal: { $ref: "#/components/schemas/Revision" }
    JobResult:
      type: object
      additionalProperties: false
      required:
        [
          result_hash,
          platform_correctness,
          strategy_evidence,
          viability,
          reproducibility,
          report_id,
          report_hash,
          confidence_label,
          research_coverage,
          disclaimer,
        ]
      properties:
        result_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        platform_correctness: { type: string }
        strategy_evidence: { type: string }
        viability:
          {
            type: string,
            enum: [undetermined, viable_for_more_research, rejected],
          }
        reproducibility: { type: string }
        report_id: { type: string }
        report_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        confidence_label:
          {
            type: string,
            enum: [local_tier_b, formal_tier_a, insufficient, rejected],
          }
        research_coverage:
          {
            type: string,
            enum: [single_run_incomplete, registered_suite_complete],
          }
        disclaimer: { type: string }
        metrics:
          type: object
          additionalProperties: { $ref: "#/components/schemas/Decimal" }
...
    RegisteredResearchReport:
      type: object
      additionalProperties: false
      required:
        [
          id,
          research_generation_id,
          manifest_hash,
          confidence_label,
          platform_correctness,
          strategy_evidence,
          viability,
          disclaimer,
          run_references,
          benchmarks,
          stress,
          capacity,
          canonical_manifest,
          created_at,
        ]
      properties:
        id: { type: string, minLength: 1 }
        research_generation_id: { type: string, minLength: 1 }
        manifest_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        confidence_label:
          { type: string, enum: [local_tier_b, formal_tier_a, rejected] }
        platform_correctness: { type: string }
        strategy_evidence: { type: string }
        viability:
          {
            type: string,
            enum: [undetermined, viable_for_more_research, rejected],
          }
        disclaimer: { type: string }
        run_references:
          type: array
          items: { type: string, minLength: 1 }
        benchmarks:
          type: array
          items: { $ref: "#/components/schemas/ResearchResultSlice" }
        stress:
          type: array
          items: { $ref: "#/components/schemas/ResearchResultSlice" }
        capacity:
          type: array
          items: { $ref: "#/components/schemas/ResearchCapacityPoint" }
        canonical_manifest: { type: string, minLength: 2, maxLength: 1048576 }
        created_at: { $ref: "#/components/schemas/Timestamp" }
    JobResource:
      type: object
      additionalProperties: false
      required: [id, kind, state, mode_label, revision, created_at]
      properties:
        id: { type: string }
        kind: { type: string, enum: [backtest, replay] }
        state:
          {
            type: string,
            enum:
              [
                QUEUED,
                RUNNING,
                PAUSE_REQUESTED,
                PAUSED,
                CANCEL_REQUESTED,
                CANCELED,
                SUCCEEDED,
                FAILED,
              ],
          }
        mode_label: { type: string, enum: [BACKTEST, REPLAY] }
        revision: { $ref: "#/components/schemas/Revision" }
        progress: { $ref: "#/components/schemas/NonnegativeDecimal" }
        cursor_ordinal: { $ref: "#/components/schemas/Revision" }
        failure_code: { type: string }
        result: { $ref: "#/components/schemas/JobResult" }
        registered_report:
          { $ref: "#/components/schemas/RegisteredResearchReport" }
        replay_inspection:
          { $ref: "#/components/schemas/ReplayEventInspection" }
        checkpoints:
          type: array
          items: { $ref: "#/components/schemas/ReplayCheckpoint" }
        input_manifest: { $ref: "#/components/schemas/LabInputManifest" }
        lifecycle: { $ref: "#/components/schemas/LabLifecycleCapabilities" }
        reproduction_bundle: { $ref: "#/components/schemas/ReproductionBundle" }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        updated_at: { $ref: "#/components/schemas/Timestamp" }
    LabInputManifest:
      type: object
      additionalProperties: false
      required:
        [
          configuration_id,
          dataset_id,
          research_generation_id,
          strategy_version,
          root_seed_hash,
        ]
      properties:
        configuration_id: { type: string, minLength: 1 }
        dataset_id: { type: string, minLength: 1 }
        research_generation_id: { type: string, minLength: 1 }
        strategy_version:
          {
            type: string,
            enum:
              [
                trend-following@1.0.0,
                mean-reversion@1.0.0,
                triangular-arbitrage@1.0.0,
                cross-exchange-arbitrage@1.0.0,
                inventory-rebalancing@1.0.0,
              ],
          }
        root_seed_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        speed: { type: string, enum: [original, accelerated, maximum] }
        incident_id: { type: string, minLength: 1 }
        first_ordinal: { $ref: "#/components/schemas/Revision" }
        last_ordinal: { $ref: "#/components/schemas/Revision" }
    LabLifecycleCapabilities:
      type: object
      additionalProperties: false
      required: [pause, resume, cancel, reproduce, compare, export]
      properties:
        pause: { type: boolean }
        resume: { type: boolean }
        cancel: { type: boolean }
        reproduce: { type: boolean }
        compare: { type: boolean }
        export: { type: boolean }
    ReproductionBundle:
      type: object
      additionalProperties: false
      required:
        [
          run_id,
          input_hash,
          manifest_hash,
          code_commit,
          go_version,
          architecture,
          operating_system,
          dataset_manifest_hash,
          dataset_revision,
          source_commit,
          configuration_hash,
          model_namespace_id,
          starting_balance_hash,
          confidence_tier,
          canonical_manifest,
        ]
      properties:
        run_id: { type: string, minLength: 1 }
        input_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        manifest_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        result_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        code_commit: { type: string, pattern: "^[0-9a-f]{40}([0-9a-f]{24})?$" }
        go_version: { type: string, minLength: 1 }
        architecture: { type: string, minLength: 1 }
        operating_system: { type: string, minLength: 1 }
        dataset_manifest_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        dataset_revision: { $ref: "#/components/schemas/Revision" }
        source_commit:
          { type: string, pattern: "^[0-9a-f]{40}([0-9a-f]{24})?$" }
        configuration_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        model_namespace_id: { type: string, minLength: 1 }
        starting_balance_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        confidence_tier: { type: string, enum: [A, B, C, D] }
        canonical_manifest: { type: string, minLength: 2, maxLength: 1048576 }
    ReplayEventInspection:
      type: object
      additionalProperties: false
      required:
        [
          event_count,
          ordinal,
          event_hash,
          canonical_event,
          canonical_decision,
          canonical_orders,
          canonical_execution_events,
          canonical_balances,
        ]
      properties:
        event_count: { $ref: "#/components/schemas/Revision" }
        ordinal: { $ref: "#/components/schemas/Revision" }
        event_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        canonical_event: { type: string, minLength: 2, maxLength: 1048576 }
        canonical_decision: { type: string, minLength: 2, maxLength: 1048576 }
        canonical_orders: { type: string, minLength: 2, maxLength: 1048576 }
        canonical_execution_events:
          { type: string, minLength: 2, maxLength: 1048576 }
        canonical_balances: { type: string, minLength: 2, maxLength: 1048576 }
    ReplayCheckpoint:
      type: object
      additionalProperties: false
      required: [revision, input_ordinal, state_hash, created_at]
      properties:
        revision: { $ref: "#/components/schemas/Revision" }
        input_ordinal: { $ref: "#/components/schemas/Revision" }
        state_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        deterministic_state_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        model_namespace_id: { type: string, minLength: 1 }
        created_at: { $ref: "#/components/schemas/Timestamp" }
    ShadowSessionRequest:
      type: object
      additionalProperties: false
      required: [configuration_id, strategy_version, portfolio_id]
      properties:
        configuration_id: { type: string }
        strategy_version: { type: string, enum: [trend-following@1.0.0] }
        portfolio_id: { type: string }
...
    ShadowSessionPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/ShadowSessionSummary" }
    ShadowDecisionSummary:
      type: object
      additionalProperties: false
      required:
        [id, outcome, reason_code, risk_outcome, risk_reason_code, occurred_at]
      properties:
        id: { type: string, minLength: 1 }
        outcome: { type: string, minLength: 1 }
        reason_code: { type: string, minLength: 1 }
        risk_outcome:
          {
            type: string,
            enum: [approved, rejected, paused, locked, not_evaluated],
          }
        risk_reason_code: { type: string, minLength: 1 }
        occurred_at: { $ref: "#/components/schemas/Timestamp" }
...
    ShadowSessionResource:
      type: object
      additionalProperties: false
      required:
        [
          id,
          state,
          label,
          public_only,
          simulation_only,
          entries_enabled,
          revision,
          created_at,
          configuration_id,
          strategy_version,
          decision_dataset_id,
          model_namespace_id,
          accepted_decisions,
          rejected_decisions,
          journal_transactions,
          activity_state,
          waiting_reason_code,
          waiting_reason,
          trigger_condition,
          input_health,
        ]
      properties:
        id: { type: string }
        state:
          {
            type: string,
            enum: [QUEUED, RUNNING, PAUSED, CANCEL_REQUESTED, CANCELED, FAILED],
          }
        label: { type: string, enum: [PUBLIC-LIVE SHADOW / VIRTUAL] }
        public_only: { type: boolean, enum: [true] }
        simulation_only: { type: boolean, enum: [true] }
        entries_enabled: { type: boolean }
        revision: { $ref: "#/components/schemas/Revision" }
        configuration_id: { type: string }
        strategy_version: { type: string }
        decision_dataset_id: { type: string }
        model_namespace_id: { type: string }
        accepted_decisions: { type: integer, minimum: 0 }
        rejected_decisions: { type: integer, minimum: 0 }
        journal_transactions: { type: integer, minimum: 0 }
        activity_state:
          {
            type: string,
            enum:
              [
                preparing,
                warming_up,
                waiting,
                evaluating,
                running,
                paused,
                blocked,
                stopped,
              ],
          }
        waiting_reason_code: { type: string, pattern: "^[a-z0-9_]{1,96}$" }
        waiting_reason: { type: string, minLength: 1, maxLength: 500 }
        next_evaluation_at: { $ref: "#/components/schemas/Timestamp" }
        trigger_condition: { type: string, minLength: 1, maxLength: 500 }
        input_health:
          {
            type: array,
            maxItems: 16,
            items: { $ref: "#/components/schemas/ShadowInputHealth" },
          }
        risk_state: { type: string, enum: [PAUSED, RESUMED, LOCKED] }
        portfolio_id: { type: string }
        run_id: { type: string }
        exchange_id: { type: string }
        slippage_model_id: { type: string }
        gap_model_id: { type: string }
        decisions:
          {
            type: array,
            items: { $ref: "#/components/schemas/ShadowDecisionSummary" },
          }
        balances:
          { type: array, items: { $ref: "#/components/schemas/ShadowBalance" } }
        positions:
          {
            type: array,
            items: { $ref: "#/components/schemas/ShadowPosition" },
          }
        pnl_attribution: { $ref: "#/components/schemas/ShadowPnlAttribution" }
        data_health: { $ref: "#/components/schemas/ShadowDataHealth" }
        orders:
          {
            type: array,
            items: { $ref: "#/components/schemas/SimulatedOrder" },
          }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        started_at: { $ref: "#/components/schemas/Timestamp" }
        stopped_at: { $ref: "#/components/schemas/Timestamp" }
        failure_code: { type: string }
...